+++ src/content/titles/endScreenTitles.ts
@@ -1,7 +1,5 @@
 import { titlesLog, titlesErrorLog } from '../../utils/logger';
 import { ProcessingResult, ElementProcessingState, TitleFetchResult } from '../../types/types';
-import { ensureIsolatedPlayer, cleanupIsolatedPlayer } from '../../utils/isolatedPlayer';
-import { currentSettings } from '../index';
 import { normalizeText } from '../../utils/text';
 import { extractVideoIdFromUrl } from '../../utils/video';
 import { fetchOriginalTitle, updateBrowsingTitleElement } from './browsingTitles';
@@ -204,9 +202,4 @@ export async function refreshEndScreenTitles(): Promise<void> {
             processingEndScreenTitle.delete(videoId);
         }
     }
-
-    // Cleanup isolated players
-    setTimeout(() => {
-        cleanupIsolatedPlayer('ynt-player-titles');
-    }, 1000);
 }